feat(protocol): 扩展消息类型字段长度并优化协议处理
- 扩展消息类型字段从1字节到2字节，修改协议头格式为15字节
- 在ByteArrUtil中添加readUnsignedShort和writeShort方法支持2字节操作
- 修改InboundMessage和OutboundMessage的消息类型字段为int类型
- 更新GameMsgType枚举值，PONG改为10001，PLAYER_JOIN_ROOM改为10002
- 优化ProtocolFrameReader中的CRC校验逻辑，改为按帧校验
- 修改ProtocolEncoder编码逻辑使用writeShort写入2字节类型字段
- 更新TcpClientManager的send方法参数类型支持2字节消息类型

+++ src/main/java/com/zunf/tankbattleclient/manager/TcpClientManager.java
@@ -80,7 +80,7 @@ public class TcpClientManager {
         startReader(in);
     }
 
-    public void send(byte type, byte version, int requestId, byte[] body) {
+    public void send(int type, byte version, int requestId, byte[] body) {
         if (!running.get()) {
             return;
         }
