+++ src/main/java/com/zunf/tankbattleclient/model/message/InboundMessage.java
@@ -2,19 +2,19 @@ package com.zunf.tankbattleclient.model.message;
 
 public class InboundMessage {
 
-    private byte type;
+    private int type;
     private byte version;
     private int requestId;
     private byte[] body;
 
-    public InboundMessage(byte type, byte version, int requestId, byte[] body) {
+    public InboundMessage(int type, byte version, int requestId, byte[] body) {
         this.type = type;
         this.version = version;
         this.requestId = requestId;
         this.body = body;
     }
 
-    public byte getType() {
+    public int getType() {
         return type;
     }
